path: introduce git_path_validate_ondisk
Introduce `git_path_validate_ondisk` to perform validations on on-disk
paths like working directory paths.  There may be system limitations on
on-disk paths, potentially ones that we enforce on a per-repository
basis.  For example: we may want to enforce (or not) MAX_PATH limitations
on Windows based on the repository's core.longpaths setting.